add-project , add-category logic

+++ src/views/components/EditableListItem.js
@@ -9,39 +9,39 @@ class ListItem extends HTMLElement {
 
    render() {
       if (this.getState()) {
-         this.innerText = this.state.title;
-         this.setAttribute('id', this.state.id);
+         this.innerText = this.state.title || 'Add New...';
+         this.setAttribute('id', this.state.id || null);
       }
       const editBtn = createElement('button')
          .setContent('edit')
-         .setAttributes({ class: 'edit' })
+         .setAttributes({ class: 'edit-item' })
          .appendTo(this);
       const deleteBtn = createElement('button')
          .setContent('X')
          .setAttributes({
-            class: 'delete',
+            class: 'delete-item',
          })
          .appendTo(this);
       const saveBtn = createElement('button')
          .setContent('save')
-         .setAttributes({ class: 'save' })
+         .setAttributes({ class: 'save-item' })
          .appendTo(this);
    }
 
    addEventListeners() {
       this.addEventListener('click', (ev) => {
-         if (ev.target.classList.contains('edit')) {
-            this.editItem();
-         } else if (ev.target.classList.contains('delete')) {
-            this.deleteItem();
-         } else if (ev.target.classList.contains('save')) {
-            this.saveItem();
-         }
+         if (ev.target.classList.contains('edit-item')) this.editItem();
+
+         if (ev.target.classList.contains('delete-item')) this.deleteItem();
+
+         if (ev.target.classList.contains('save-item')) this.saveItem();
       });
    }
 
    editItem() {
       this.contentEditable = true;
+      this.querySelector('.edit-item').style.display = 'none';
+      this.focus();
    }
 
    deleteItem() {
@@ -53,10 +53,14 @@ class ListItem extends HTMLElement {
    }
 
    saveItem() {
-      pubsub.publish(
-         `${this.getAttribute('data-type')}:save`,
-         this.getAttribute('id')
-      );
+      if (this.getAttribute('id')) {
+         pubsub.publish(
+            `${this.getAttribute('data-type')}:save`,
+            this.getAttribute('id')
+         );
+      } else {
+         pubsub.publish(`${this.getAttribute('data-type')}:add`);
+      }
    }
 }
 customElements.define('editable-li', ListItem);
